* thread.c : add Mutex#sleep ability to accept block to run before lock
To fix case with ConditionVariable#wait timeout, simplest way is to
remove thread from sleepers before acquiring lock on mutex. To do
it, we should extend Mutex#sleep method to have ability accept block
and run it before trying to lock.
Perhaps, it is not the best way to do thing, but the simplest one.